Refactor test_divide-related strategies
