+++ tests/strategies/geometry/base.py
@@ -1,5 +1,8 @@
+from fractions import Fraction
+from math import floor
 from numbers import Real
-from typing import TypeVar
+from typing import (List,
+                    TypeVar)
 
 from gon.discrete import Multipoint
 from gon.linear import Contour
@@ -14,19 +17,27 @@ from tests.pode_tests.config import (MAX_CONTOUR_SIZE,
                                      MAX_HOLES_SIZE,
                                      MIN_CONTOUR_SIZE,
                                      MIN_COORDINATE,
-                                     MIN_HOLES_SIZE,
-                                     coordinates_strategies_factories)
+                                     MIN_HOLES_SIZE)
 from pode.utils import joined_constrained_delaunay_triangles
 
 T = TypeVar('T')
 
 TRIANGULAR_CONTOUR_SIZE = 3
 RECTANGLE_CONTOUR_SIZE = 4
+MIN_PARTITION_SIZE = 1
 
 fractions = st.fractions(MIN_COORDINATE, MAX_COORDINATE)
 raw_fraction_contours = planar.contours(fractions)
 fractions_contours = st.builds(Contour.from_raw, raw_fraction_contours)
+fraction_triangles = st.builds(Polygon.from_raw,
+                               planar.polygons(fractions,
+                                               max_size=TRIANGULAR_CONTOUR_SIZE))
+convex_divisors = st.sampled_from([constrained_delaunay_triangles,
+                                   joined_constrained_delaunay_triangles])
 
+coordinates_strategies_factories = {int: st.integers,
+                                    Fraction: st.fractions,
+                                    float: st.floats}
 coordinates_strategies = st.sampled_from([
     factory(MIN_COORDINATE, MAX_COORDINATE)
     for factory in coordinates_strategies_factories.values()])
@@ -42,27 +53,71 @@ def coordinates_to_polygons(coordinates: st.SearchStrategy[Real]
         max_holes_size=MAX_HOLES_SIZE))
 
 
-def coordinates_to_triangles(coordinates: st.SearchStrategy[Real]
-                             ) -> st.SearchStrategy[Polygon]:
-    return st.builds(Polygon.from_raw, planar.polygons(
-        coordinates,
-        max_size=TRIANGULAR_CONTOUR_SIZE))
-
-
 def coordinates_to_multipoints(coordinates: st.SearchStrategy[Real]
-                               ) -> st.SearchStrategy[Polygon]:
+                               ) -> st.SearchStrategy[Multipoint]:
     return st.builds(Multipoint.from_raw, planar.multipoints(coordinates,
                                                              min_size=1))
 
 
 polygons = coordinates_strategies.flatmap(coordinates_to_polygons)
-fraction_triangles = st.builds(Polygon.from_raw, planar.polygons(
-    fractions, max_size=TRIANGULAR_CONTOUR_SIZE))
 multipoints = coordinates_strategies.flatmap(coordinates_to_multipoints)
-_fraction_points = st.builds(Point, fractions, fractions)
-unique_points_pairs = st.builds(tuple,
-                                st.sets(_fraction_points,
-                                        min_size=2,
-                                        max_size=2))
-convex_divisors = st.sampled_from([constrained_delaunay_triangles,
-                                   joined_constrained_delaunay_triangles])
+
+
+def requirements(sum_: Real,
+                 *,
+                 min_value: Real = 0,
+                 size: int = MIN_PARTITION_SIZE,
+                 base: st.SearchStrategy[Real] = st.integers()
+                 ) -> st.SearchStrategy[List[Real]]:
+    if size < MIN_PARTITION_SIZE:
+        raise ValueError('`size` should not be less '
+                         f'than {MIN_PARTITION_SIZE}.')
+    if not (0 <= min_value <= sum_):
+        raise ValueError(f'`min_value` should be in [0, {sum_}] interval.')
+    if min_value:
+        max_size = floor(sum_ / min_value)
+        if max_size < size:
+            raise ValueError(f'`size` should not be greater than {max_size}.')
+
+    def to_proportions(numbers: List[Real]) -> List[Real]:
+        return [2 * abs(number) / (1 + number * number) for number in numbers]
+
+    def to_partition(proportions: List[Real]) -> List[Real]:
+        factor = sum_ / sum(proportions)
+        return [proportion * factor for proportion in proportions]
+
+    def bound_minimum(partition: List[Real]) -> List[Real]:
+        minimum = min(partition)
+        if minimum >= min_value:
+            return partition
+        partition_size = len(partition)
+        denominator = sum_ - partition_size * minimum
+        slope = sum_ - partition_size * min_value
+        intercept = sum_ * (min_value - minimum)
+        return [max((part * slope + intercept) / denominator, min_value)
+                for part in partition]
+
+    def normalize(partition: List[Real]) -> List[Real]:
+        partition_sum = sum(partition)
+        if partition_sum < sum_:
+            arg_min = min(range(len(partition)),
+                          key=partition.__getitem__)
+            partition[arg_min] += sum_ - partition_sum
+        elif partition_sum > sum_:
+            arg_max = max(range(len(partition)),
+                          key=partition.__getitem__)
+            partition[arg_max] -= partition_sum - sum_
+        return partition
+
+    def is_valid(partition: List[Real]) -> bool:
+        return sum(partition) == sum_
+
+    return (st.lists(base,
+                     min_size=size,
+                     max_size=size)
+            .filter(any)
+            .map(to_proportions)
+            .map(to_partition)
+            .map(bound_minimum)
+            .map(normalize)
+            .filter(is_valid))
